feat(nodejs): Make global npm packages configurable per host
This commit introduces a new variable 'globalNpmPackages' in the
host-specific variables file to allow for different sets of globally
installed npm packages on each host.

The 'nodejs' module is updated to use this new variable.

Additionally, the 'claude-code' npm package is updated to the latest
version and a hint for getting the SHA256 hash for a file has been added
to the README.md.

Also installing the `gemini-cli` as free alternative to `claude-code`