refactor(nodejs): Use a list for global npm packages
This commit refactors the 'globalNpmPackages' variable to be a list of strings instead of a single space-separated string. This makes the list of packages easier to read and manage.

The 'nodejs' module is updated to use 'lib.concatStringsSep' to join the list into a single space-separated string for the 'volta' command.